Changed model for wordlists to allow for really large sizes

+++ hashview/models.py
@@ -110,7 +110,7 @@ class Wordlists(db.Model):
     owner_id = db.Column(db.Integer, db.ForeignKey('users.id'), nullable=False)
     type = db.Column(db.String(7))                          # Dynamic or Static
     path = db.Column(db.String(245), nullable=False)
-    size = db.Column(db.Integer, nullable=False)
+    size = db.Column(db.BigInteger, nullable=False)
     checksum = db.Column(db.String(64), nullable=False)
 
 class Tasks(db.Model):
